Move identifier type to new module
