+++ dj_link/entities/custom_types.py
@@ -1,0 +1,4 @@
+"""Contains custom types."""
+from typing import NewType
+
+Identifier = NewType("Identifier", str)
