+++ dj_link/use_cases/pull.py
@@ -5,8 +5,9 @@ import logging
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, List, Set
 
+from ..entities.custom_types import Identifier
 from ..entities.link import create_link, pull
-from ..entities.state import Components, Identifier
+from ..entities.state import Components
 from .base import AbstractRequestModel, AbstractResponseModel, AbstractUseCase
 
 if TYPE_CHECKING:
